move the python test
